+++ assets/javascripts/kitten/components/buttons/button/test.js
@@ -224,4 +224,32 @@ describe('<Button />', () => {
       expect(component).toMatchSnapshot()
     })
   })
+
+  describe('with variant', () => {
+    beforeEach(() => {
+      component = renderer
+        .create(<Button variant="orion">Orion</Button>)
+        .toJSON()
+    })
+
+    it('matches with snapshot', () => {
+      expect(component).toMatchSnapshot()
+    })
+  })
+
+  describe('with variant and fluid', () => {
+    beforeEach(() => {
+      component = renderer
+        .create(
+          <Button variant="orion" fluid>
+            Orion fluid
+          </Button>,
+        )
+        .toJSON()
+    })
+
+    it('matches with snapshot', () => {
+      expect(component).toMatchSnapshot()
+    })
+  })
 })